refactor: 重構 Carousel 元件，簡化結構並整合至首頁，提升食譜展示效果與使用者互動性。
