+++ src/pages/index.tsx
@@ -1,14 +1,13 @@
 import type React from 'react';
 import { useState } from 'react';
-import { Plus, ChevronDown, X, ChevronRight, ChevronLeft } from 'lucide-react';
+import { Plus, ChevronDown, X } from 'lucide-react';
 import { Header } from '@/components/layout/Header';
 import { Footer } from '@/components/layout/Footer';
 import { Button } from '@/components/ui/button';
 import { Tabs, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { CategoryCard } from '@/components/ui/CategoryCard';
 import { RecipeCard } from '@/components/ui/RecipeCard';
-import useEmblaCarousel from 'embla-carousel-react';
-import { cva } from 'class-variance-authority';
+import { Carousel } from '@/components/ui/carousel';
 
 // 定義食譜類型
 type Recipe = {
@@ -29,107 +28,6 @@ type Category = {
   image: string;
   description: string;
 };
-
-// 定義輪播容器樣式
-const carouselContainerVariants = cva('relative group', {
-  variants: {
-    size: {
-      default: 'w-full',
-      compact: 'max-w-screen-lg mx-auto',
-    },
-  },
-  defaultVariants: {
-    size: 'default',
-  },
-});
-
-// 定義輪播軌道樣式
-const carouselTrackVariants = cva('flex', {
-  variants: {
-    spacing: {
-      default: '-ml-4',
-      tight: '-ml-2',
-    },
-  },
-  defaultVariants: {
-    spacing: 'default',
-  },
-});
-
-// 定義輪播導航按鈕樣式
-const carouselButtonVariants = cva(
-  'absolute top-1/2 -translate-y-1/2 z-10 bg-black/50 text-white rounded-full p-2 opacity-0 group-hover:opacity-100 transition-opacity',
-  {
-    variants: {
-      position: {
-        left: 'left-0',
-        right: 'right-0',
-      },
-    },
-    defaultVariants: {
-      position: 'left',
-    },
-  },
-);
-
-// Netflix風格輪播元件
-function CarouselSection({
-  title,
-  items,
-  renderItem,
-}: {
-  title: string;
-  items: any[];
-  renderItem: (item: any) => React.ReactNode;
-}) {
-  const [emblaRef, emblaApi] = useEmblaCarousel({
-    align: 'start',
-    loop: false,
-    skipSnaps: false,
-    dragFree: true,
-  });
-
-  // 處理向前滾動
-  const scrollPrev = () => {
-    if (emblaApi) emblaApi.scrollPrev();
-  };
-
-  // 處理向後滾動
-  const scrollNext = () => {
-    if (emblaApi) emblaApi.scrollNext();
-  };
-
-  return (
-    <div className={carouselContainerVariants()}>
-      <h2 className="text-xl font-bold mb-3 px-4">{title}</h2>
-
-      {/* 前滾按鈕 */}
-      <button
-        className={carouselButtonVariants({ position: 'left' })}
-        onClick={scrollPrev}
-        aria-label="查看上一頁"
-      >
-        <ChevronLeft size={24} />
-      </button>
-
-      {/* 輪播區域 */}
-      <div className="overflow-hidden" ref={emblaRef}>
-        <div className={carouselTrackVariants()}>
-          {items.map((item) => renderItem(item))}
-        </div>
-      </div>
-
-      {/* 後滾按鈕 */}
-      <button
-        className={carouselButtonVariants({ position: 'right' })}
-        onClick={scrollNext}
-        aria-label="查看下一頁"
-      >
-        <ChevronRight size={24} />
-      </button>
-    </div>
-  );
-}
 
 /**
  * 網站首頁組件
@@ -315,10 +213,9 @@ export default function HomePage() {
           <div className="text-gray-400">圖片區域</div>
         </div>
 
-        {/* 季節食譜區塊 */}
         {/* 季節食譜區塊 - Netflix風格輪播 */}
         <section className="py-3">
-          <CarouselSection
+          <Carousel
             title="季節食譜"
             items={seasonalCategories}
             renderItem={(category) => (
@@ -329,7 +226,7 @@ export default function HomePage() {
 
         {/* 特色食譜區塊 */}
         <section className="py-3">
-          <CarouselSection
+          <Carousel
             title="特色食譜"
             items={specialCategories}
             renderItem={(category) => (
